+++ tests/data/general/correct/numpy_model_field.py
@@ -1,0 +1,18 @@
+"""
+When used as a model field, checks against a numpy array
+"""
+
+from typing import Any
+from typing import Literal as L
+
+import numpy as np
+from pydantic import BaseModel
+
+from numpydantic import NDArray, Shape
+
+
+class Foo(BaseModel):
+    x: NDArray[Shape[L[3], L[3], L[3]], Any]
+
+
+foo = Foo(x=np.zeros((3, 3, 3)))