+++ tests/data/general/correct/shape_args.py
@@ -1,0 +1,7 @@
+from typing import Any
+
+import numpy as np
+
+from numpydantic import NDArray, Shape
+
+x: NDArray[Shape[1, 2, 3], Any] = np.zeros((1, 2, 3))